Add macro functions, plus example text.
Macro functions work much like C's preprocessor macro functions, but
with some tweaks since this language uses curly-braces for
macro-expansion instead of expanding bare words. The defining syntax is
 #macro(arg1, arg2, ...) body
while using them is {macro(arg1,arg2,...)}. (Whitespace is not
significant in the argument list when defining a macro function, but it
*is* significant when calling them.)

Parenthesis must be balanced inside a macro call, and commas only
separate arguments at the outermost nesting level. This allows for
things like {macro(if(condition, 0, 1), "error msg")} to work as
expected, at the cost of disallowing things like {macro("(")}. (Using
extra macro trickery can achieve the latter if it's really needed.)

This also expands the syntax to allow for nesting macro calls, even for
non-functions - i.e. {foo{bar}}. Technically that by itself gives a
great deal of additional power, but not as much as functions give.

This changes the evaluation of macros to be evaluated at call time
instead of define time. This doesn't change the semantics of any
previously valid programs, but does allow for newly valid ones where
macros can refer to macros that aren't defined yet. It also means you
can have infinite recursion - this is guarded with a counter.