Refactor Lua code so that it doesn't access javascript.
This moves the API boundary a bit so that runLua() is the interface
point between javascript and lua now, and is the primary place on the JS
side that needs to know about lua. The lua side becomes completely
unaware of JS existing. This makes things more portable, as well as
enabling future improvements.