Add the ability to backslash-continue lines.
This is done very early in parsing, so that it works for both regular
lines and macro definitions. This means it also continues comments, with
the unintuitive result that the comment continues onto the next line.

This should not break any existing programs, because backslash was never
valid at the end of a line before - it would have only been legal inside
strings, which couldn't wrap lines.