Move all Lua processing to a Web Worker.
This unblocks the main event loop, so that typing and other interactions
don't get stalled all the time. As a side effect, loading is much
faster, because running the unittests no longer blocks startup of
everything. (It does block the first compile results, but that's not
nearly as noticable.)

This also changes script saving to happen when compiles come back, so
that the "saved" prompt also signals when the compile window is
up-to-date. This also means saves are much more responsive, as well.