Add :import to the language.
This is a massive change: Many things had to be reworked, and I wouldn't
be surprised if there's still some bugs. I also added/changed how the
unit-tests work, so that tests can be added that aren't just
import/export tests. (Needed for features that exist only in source-code
form!)