Take running the unittests out of startup.
Instead, the tests are triggered with an easter-egg sequence of trying
to rename the Default Workspace.
This removes them from the critical path, saving multiple seconds from
startup. It also makes it easier to tell when they've explicitly passed.

It's still not as good as a real test harness, but moving closer to the
right direction.