Update name display for upcoming package changes.
With the new package system, there are two length limits to worry about
instead of just one. Although we never actually limit the length of
names in exported scripts (we let the game do that, in case our logic is
wrong,) we want the display to mirror the game's logic.