Fix issue with a.d() and a.i()
When they had a string constant as their operator argument, but the
string wasn't a recognized operator, on import the string would get
unquoted, leading to a syntax error.

Even though the code in question won't *run* successfully, it still
should compile properly (and round-trip properly). Especially since we
might be dealing with code from a future game version that has a new
operator.

Also add tests for this and the Inf/NaN parsing issues.