Add the ability to export/import workspaces as source.
This saves the whole workspace as a JSON blob, including the name.
Import is accomplished using the same import button as for AI export
code (the formats are distinguishable).

This also improves the workspace deletion handling, since it was badly
needed for testing. Deleting a workspace deletes all its scripts, and
deleting "All" deletes everything.