Big compilation refactor
This moves a chunk of the formatting logic out of Lua and into JS.
Lua's compile() now returns a data structure and it is up to the JS to
organize this into a string to put in the text box.

So far this just preserves the existing functionality (with some minor
changes like having summary stats for workspace exports). Supporting
compression comes next.