Revert "Add validator/help messages for arithmetic and comparison functions."
This reverts commit 4e80eec67ddddcc182efc0d9d2fe2990ad67aec1.

It broke importing scripts, which was bad. But, it can't even be
salvaged because it unavoidably causes scripts with an unrecognized
operator to fail to compile (this is the entire point). We are
explicitly testing that this works in the unittests, because new
operators can be added in the future.

Given that direct use of arithmetic.x and comparison.x are advanced
features, I think it is fine for them to not be checked.

This revert *does* leave the typo-fix in. And it bumps the version to
1.86 so everyone knows they're on the fixed version.